priemerny cas v prevadkze

+++ diss/_1proj/src/main/java/org/main/shared/Statistics/AverageVehiclesInSTK.java
@@ -8,7 +8,7 @@ public class AverageVehiclesInSTK extends Statistics {
         if (vehicleLeft > vehicleArrived) {
             System.out.println("warning: more vehicles left than arrived");
         }
-        return vehicleArrived - vehicleLeft;
+        return (double)vehicleArrived - vehicleLeft;
     }
 
     @Override
